+++ qiskit/providers/ibmq/utils/validators.py
@@ -1,0 +1,42 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2019.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Custom fields for validation."""
+
+from marshmallow.exceptions import ValidationError
+
+from qiskit.validation import ModelTypeValidator
+
+
+class EnumType(ModelTypeValidator):
+    """Field for enums."""
+
+    def __init__(self, enum_cls, *args, **kwargs):
+        self.valid_types = (str, enum_cls)
+        self.valid_strs = [elem.value for elem in enum_cls]
+        self.enum_cls = enum_cls
+        super().__init__(*args, **kwargs)
+
+    def _serialize(self, value, attr, obj):
+        return value.value
+
+    def _deserialize(self, value, attr, data):
+        return self.enum_cls(value)
+
+    def check_type(self, value, attr, data):
+        # Quick check of the type
+        super().check_type(value, attr, data)
+
+        if (isinstance(value, str)) and (value not in self.valid_strs):
+            raise ValidationError("{} is {}, which is not an expected value.".format(attr, value))
